+++ __tests__/cases/server/server.js
@@ -38,6 +38,8 @@ const CONFIG_FIELDS = [
 	'autoprefixer',
 	'babelTranspileModules',
 	'cookies',
+	'createMissingComponents',
+	'generatedComponentsExtension',
 ];
 
 describe('Server instance', () => {